Introduce basic data model for ticket
The ticket is a piece of information that has some unique identifier (ID) and
contains some data (Body).

In typical issue tracker there will be as well some crutial metadata at
disposal, such as a reporter, the time of creation/modifcation, coeditor etc.
The idea behind this model is that any kind of history related to the ticket
is being handled by git system, not an issuer itself. The history of the ticket
should be available to potential viewer/editor no matter the tool they use.

The same claim can be said about the exact content of the ticket. It is
expected to be human readable - hence I assume  to use there a markdown
file format.