Add README.md file
This is a first section of the README.md file explaining some basic
ideas behind the project. The manifesto is in it's draft form and could
be altered in future. At this point I decided to make use of a TODO.md
as a support for tasks arrangement and adding some workflow metadata.
There are no assumptions on configuration nor the way of storing issues.