Add a markdown support for issue build
This commit introduces the markdown support for issue title and body.

The issue is now supposed to be a markdown document, and a title for
this issue is the text of the first markdown node of that document. So
for instance when the document is just a plain text then the first
paragraph is considered a title. If the document is formatted and on the
very top someone puts a header then this header is consideres a title.

The identificator of the issue is a title, stripped of the non
alphanumeric characters and words concatenated tith the hyphen.