Allow message related commands to skip -m option
In case of the following commands:
* git issues report
* git issues refine
the `-m`/`--message` option is now optional. The lack of that option
indicates that the command should take the message from the editor, not
a CLI parameter.

In case of `git issues list` it's `QUERY` option is considered optional.
The absence of the query indicates the full list of tickets to be shown
in return.